Refactors Zip4 Query Tool into two discrete tools
- Extends single use case scenario to generate a comparison between Smarty and USPS counties for a single ZIP code into one script zip-to-county-diff-single

- Combines the functionality of countylookup and smarty-usps-comparison into one tool, zip-to-county-diff-report

- Exports some functionality of the cmd smarty-usps-comparison so that both zip-to-county-diff-single and zip-to-county-diff-report executables have shared access to them.

Tested: Test ran both zip-to-county-diff tools and observed the output was as expected.
