--- cmd/countylookup/.gitignore
@@ -1,1 +1,0 @@
-countylookup
