--- cmd/smarty-usps-comparison/.gitignore
@@ -1,1 +1,0 @@
-smarty-usps-comparison
